refactor: structure part
